Added section header parsing

+++ src/program.c
@@ -40,7 +40,7 @@ get_program_headers(FILE *fp, uint16_t e_phnum)
 	for (int i = 0; i < e_phnum; i++)
 	{
 		hdrs[i] = (program_header *)malloc(sizeof(program_header) + 1);
-		for (int j = 0; j < sizeof(program_header); j++)
+		for (size_t j = 0; j < sizeof(program_header); j++)
 			((char *)(hdrs[i]))[j] = fgetc(fp);
 	}
 	print_program_headers(hdrs, e_phnum);
